Performance boost for LMFitModel by roughly a factor of 6-7. Performance of LMFitModel now beats the level lmfitting in kMap.py had before the introduction of the LMFitModel, however, not quite as fast as the example5_old.py script (takes about 1.2-3 times as long; depending on how it is measured).
Lmfit has the tendency to not explore the phasespace for a parameter if the parameter starts at the boundary. This means a background fit with an initial value c=0 will not yield the correct results (and it takes longer) because lmfit does not try any other value besides 0. This behaviour is present in the example5_old.py script, the new LMFitModel as well as the old lmfitting in kMap.py.