Bug fix for Issue 7:
Closing an options window for the database windows before closing the database window itself caused a Runtime Error because it tried to close the already closed window.

This should not be possible as the close button for the options windows has been specifically disabled to prevent this (at least on Linux they are).
The issue has been addressed by catching the Runtime error.