Bug fix for Issue 9:
Closing the mainwindow did not close any sub windows like MatplotLib windows or database browsers.

The problem was that the mainwindow did not assume ownership of these sub windows, however, the subwindows are in qt jargon mainwindows and as such keep the program running.

The issue was addressed by giving the mainwindow ownership of those windows and catching the closing event. Then manually closing of them still open.

Many thanks to Xiaosheng for reporting the bug and providing the solution.