Added the database ID to orbital metadata. Addresses issue 3.
For orbitals loaded through the database the database ID will be added
to the meta data and thus displayed in the tooltip.
The database ID is construct as such: 'molecule_ID'.'orbital_ID'.
(For example: the orbital 12 in molecule 253 has the ID: 253.12)

Unfortunately there exists a name conflict for 'ID'. The number in the 'ID'
column in the database window (=database_ID) has nothing to do with the
ID throughout the rest of the program and each piece of data will have
one of the latter, but only orbitals loaded through the database will
have the former.