Added the option to swap axes. Addresses issue 13.
There is a new button labeled 'Swap axes' on every data slider. It will
swap the axes of the slice data loaded in the tab which are not the
slice axes (e.g. if 'E_kin' slices than it will swap 'kx' and 'ky').

Please note that this will swap the axes for all tabs this data object
is loaded.
Works on the lmfit tabs as well.