New Feature: Fittable Background Equation. Addresses issue 5.
This update brings fittable background equations. The line edit on the
lmfit tab now takes an arbitrary number of variables (see tooltip for
more information). Those variables get added automatically to the
tree and can be used or even fitted.

Additionally fixed example scripts.

ATTENTION: There is a rather rare bug (different size return value to
lmfit) that was not reproducable and thus fixable so far.