Added a MANIFEST.in and simplified paths in scripts. Addresses issue 14.
There has been a long standing problem with kmap imports outside the
package. This issue was most prominent in the example scripts regularly
producing a missing key error ('cube').

The reason for this has been identified and fixed. In addition the
paths in the example scripts are handled more elegantly.
The new way is to be used program-wide as well sometime in the future.