Added "Slice Like Plots" on ProfilePlotTab (Issue 22).
Add the "Slice Like Plots" radio button on the ProfilePlotTab. It will
only appear for SlicedDataTabs and if selected all crosshair regions
including the 'center' option will become available. The latter is only
for "Slice Like Plots" available.
Additionally, the sample spinboxes will disappear as they are not needed
for this type of plot.

The plot will be the sum (or mean, depends on settings) of the region
chosen for each slice. If 'Normalized' is off, it's basically like
sliding through the data and entering the values the crosshair displays
into a plot.