Added a new setting for the colormap export. (issue 28)
Added the nPts setting in the general settings file. It determines how many color points get exported from the histogram widget when requesting a LUT. This is important for the matplotlib window as it will use this LUT. Default is 500. Changing this value will change the display of the image in the matplotlib window.