Custom colormap and equation files can now be used. (issue 26)
Added the possibility to add a second colormap/background_equation file to the same folder ending with "_user" (colormaps_user.json or background_equations_user) which will be preferably used if they exist. _default files are fallback option. The content of both files will NOT be merged. If a "_user" file is present it and only it will be used.

Added also the option to change the search path for these files in the settings file.