Export to hdf5 for OrbitalDataTab (issue 15)
Added the option to export to hdf5 for the OrbitalDataTab. For this
reason the new menu item "File -> Export -> Hdf5" (default shortcut:
Ctrl+h) was introduced.
The data format is such that it can be loaded by kMap.py as SlicedData
again. For this reason at least two different kinetic energy kmaps need
to be exported. In general settings file there is a new setting under
"orbital" where you can specify the energy slices that get exported.
Other settings like polarisation will be carried over as well.
Please note that the interpolation has to be activated!