"Match" button for kinetic energy on OrbitalDataTabs (issue 32)
Reintroduction of the "Match" button in the "Cube Options" (formerly
known as "Other Options") on OrbitalDataTabs. Clicking it will not do
anything if no SlicedDataTab is opened. If one or more are open the
first one will be picked and either the average, the max or the min of
the currently as slice axis declared axis will be entered as kinetic
energy on the OrbitalDataTab. Additionally it can be configured to
simply use the current slice value. To configure please see the new
"match_type" setting in the general settings under "orbital".