Dark Theme and Saving Path for Matplotlib (issue 35 & 36)
Added the option (see settings: [app]) to activate a dark theme. It is
a custom theme in the metallic style. It looks nice but messes with the
layout and thus can not be recommend in general. If you like it and would
want to use it if it weren't for the layout isses, tell me and I will try
to fix them if the demand is reasonable.

Added a new path in the settings to set the default path for Matplotlib
images.