Two bug fixes (IMPORTANT NOTE BELOW; issue 39 & 40)
IMPORTANT: Old save files that were invalid due to issue 40 are still
invalid as the problem was in the saving process not the loading process.
If you have an old save file you would like to open PLEASE contact me and I
will update the save file or include a special case in the loading process
to make it happen.

Issue 39: The case of not acutally loading a file (and thus an empty
list of file names) was not properly caught and provoked a crash. Fixed
by explicitly catching this case.

Issue 40: Orbitals loaded from the database had their ID not saved in
the list of orbital properties as was the case for orbitals loaded
from disk. Thus the length of properties was different and at the time
of loading a project file the properties list is unfolded by the star
operator (bad habit/coding).
Added the ID to the properties. Old files are still not valid but new
ones are now.