Added the "None" option to paths (issue 37)
Some of the paths (cube_start, hdf5_start, project_start,
hdf5_export_start, matplotlib) in the settings file now have the option
to be set to "None". This means that no path will be set and the OS takes
over. In general you will land inside the project folder or the root
directory of your OS the first time you trigger a FileDialog in the
current session. However, from the second time on the OS remembers the
previously used path and you will start there. Useful if the particular
directory is session depended.

Some caveats though:
1) This does not work between sessions. If you reboot the program the OS
will not have remembered the desired path.
2) The OS only remembers one path per program. This means, even if you
have multiple paths (for example hdf5_start and cube_start) set to None,
the OS will remember the last directory, regardless whether it was when
loading hdf5 files or cube files.
3) You do not have to actually save a file for the OS to remember. Triggering
a dialog is enough.
4) The matplotlib option does only remember after actually saving file.