Added export to matplotlib for LMFit Tabs (issue 42)
The residual plot and the sum plot (the last two) on the LMFit and the
LMFitResult Tab are now exportable to matplotlib. Due to the way this
feature is implemented, both windows will open at the same time, everytime.
There is no option to only show one of them for now.
The first two plots (sliced data and orbital data) do not get exported
as they show nothing not already displayable in the already open SlicedDataTab
or OrbitalDataTab respectively.

The LUT issue 43 is not resolved yet.