Introduction of the "s_share" setting and a bug fix (issues 47 & 49).
As a further fix of issue 47 the s_share setting that was implemented
during the last commit for the "get_kmap" method is now program-wide
available. It has been implemented as setting under the "[orbital]"
group in the general settings file.
This setting will now be respected on the orbital data tab, the sliced
data tab for sliced data generated from hnu dependence and from orbitals,
on the lmfit and the lmfit result tab.
For users of the lmfit_model.py without gui there is now also the option
to change value. It is initialized with a default of 0.694.

The unpolarized He-lamp option has been missing on the lmfit tab (issue
49). It has now been included.