Residuals can now be exported to hdf5 (issue 57).
When on a LMFitResultTab (not a regular LMFitTab!) using the "File ->
Export -> Hdf5" will export the residual data for all slices as a
hdf5 file in the format necessary to be loaded in the SlicedDataTab.

ATTENTION: SlicedData needs to have at least two slices. Using the
'Only this Slice' option when fitting will still let you export the
data but it will NOT be loadable! Use the 'All Slices individually' if
you want to load the data again.