Matrix Inversion method to the lmfit tab added (issue 64)
Added the option to do a deconvolution by solving the system of linear equations resulting from a least square fitting the weights and a background. The system is solved by matrix inversion and only allows deconvolution as other fitting parameters would result in a non linear problem.

The main advantage is speed as the system has only a low number of unknown variables (# of orbitals + 1 background) and thus the matrix inversion is basically instant. The speed gain over lmfit increases drastically with number of orbitals.

The use should not differ from other lmfit methods except for two points:
1. Fitting two identically orbitals will not yield results as this would result
in a singular matrix. Two orbitals with different angles is possible.
2. The background equation has have a mulitplicative constant c in front. Background maps can still take any non constant form (e.g. Gaussian) but only the constant c will can be fitted.