LMFit model will now return cut kmaps for other than residual as well (issue 63)
The lmfit model used to only cut the ROI or annulus at the very end for the residual map. The others (sliced, orbital, weighted sum) were returned whole.
The change was made so that all maps on the lmfittab now only show the roi.

Additionally changing the crosshair inside the map via handles now triggers a redraw.