Medium-sized update to the crosshair system.
The crosshair has a new column with information about the total
(non-NaN) area, the (non-NaN) area inside the ROI and the (non-NaN) area
inside the annulus in area units provided by the loaded data (e.g. if
the x- and y- axis are in 1/A the area displayed has the units of
1/A^2).
The intensity column can either display the integral of the intensity
inside the corresponding area (total, point, ROI or annulus) or the
normalized intensity which is equal to the integrated density divided by
the area. The behaviour can be changed via the general config file using the
"crosshair.normalized_intensity" setting ("True" or "False"). In case
of "False" the intensity will have the units of the data, in case of
"True" it will be the units of the data / units of area (explained
above; e.g. a.u.*A^2)
The default is normalized.

The intensity labels on the orbitaldatatab will be updated
much more frequently now.

The display of area and intensity have been switched to scientific
notation (e.g. 3e1 = 30). The number of decimal places (digits behind
the .) can be customized via the "crosshair.decimal_places" setting in
the general config file.