Added partial fits.
Added the ability to fit only a certain slice range at a time.
For this purpose there is a new option for the slice policy called "Some
Slices". Choosing it will enable two new spinboxes to the right where
the range of slices to be fitted can be chosen.

This update is the first part.