Update to partial fits.
Second part of the update to this feature.
The LMFitTab (not the LMFitResultTab!) now remembers old results and
only updates the results for those slices that have been newly fit. E.g.
if the slices 2 to 10 are fitted and then, on the same LMFitTab the
slices 5 to 12 are fitted, the second LMFitResultTab will have results
for the slices 2 to 12, where the slices 2-4 are the first results and 5
to 12 are the new ones.

This feature will most likely not work with the save project feature at
all. The save project feature will have to be reworked from scratch
anyhow in the near future.