Added cached cubefile feature.
This update introduces a new setting general_settings.paths.cache which
takes a path on your local machine. If this path is set kMap will try to
find any cubefile chosen to be downloaded from the online database in
this path first and if found load it instead of downloading anything. If
this path is not valid, set to 'None' or the corresponding file is not
found there kMap will attempt to download it as before.
If the path is valid but the file was not found it will be stored there
automatically for the next time it is loaded.

To disable this feature simply do not enter a valid path for the cache
setting.