Bundled release of several new features (v.2.0.0b)
This version is seen as a bundle of all changes since bc97a97 and is
labeled v2.0 because of major incompatabilities regarding the "Save
Project" feature (see below).

Summary of changes:
- Complete overhaul of the "Save Project" Feature (see below).
- Partial Fits (see 5401223 & 9144114)
- Cache for downloaded files (see 13b99d3)
- Update to the crosshair (see bc97a97)
- Change in logging persistence setting

Saving Project Overhaul:
Biggest change in this bundle is the complete overhaul of the "Save
Project" feature. Thus savefiles from previous versions are NOT
compatible with future ones.
Savefiles are now transferable between different computers and
installations of kMap. Data will not be saved inside the savefile and
has to be transfered by other means. Loading a savefile will look for
locally loaded files in a number of places. The files cannot change name
but can change location. To add a search path for files add it to the
new general_settings.paths.path variable (comma separated).